stockholm/requirements.txt
altgraph==0.17.4
packaging==24.0
pybase64==1.3.2
pycryptodome==3.20.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
